Initial Commit with basic plots
Added the read_data function to read data from a CSV file and output it to a 2d list.

Added the xy_graph function which plots the data in an xy graph from the read_data function

Added the log_graph function which plots the data after taking logs to base 10 of both axes.

Added a basic data set of 50 data points to test the