Basic Linear Fits
Added numpy and SKLearn dependencies

Added r_squared() function, which calculates the r^2 value for a linear fit using the SkLearn r2_score function.

Added a linear_fit function, which utilises the numpy polyfit function to fit a linear curve to the data.

Added annotations to the log graph to display the r^2 value on the graph itself.

Added basic function to 2 data sets at the same time [TEMPORARY].

Changed the log function, so the log of the data is taken before the data is plotted for linear fitting purposes.

Changed the formatting of the plots so that each data point is not connected.